requirements.txt
tweepy==3.7.0
kafka-python==1.4.4
pandas==0.24.1
Flask==0.10.1
